requirements_windows.txt
appdirs==1.4.4
argcomplete==3.5.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==23.0.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
python-barcode==0.15.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.0
setuptools==75.3.0
six==1.16.0
Werkzeug==3.1.1
